requirements.txt
Django==3.2.11
PyJWT==1.7.1
XlsxWriter==3.0.3
adal==1.2.7
asgiref==3.4.1
azure-common==1.1.28
azure-core==1.24.1
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-network==2.7.0
azure-mgmt-notificationhubs	==2.1.0
azure-mgmt-nspkg	==3.0.2
azure-mgmt-policyinsights	==0.1.0
azure-mgmt-powerbiembedded	==2.0.0
azure-mgmt-rdbms	==1.9.0
azure-mgmt-recoveryservices	==0.3.0
azure-mgmt-recoveryservicesbackup	==0.3.0
azure-mgmt-redis	==5.0.0
azure-mgmt-relay	==0.1.0
azure-mgmt-reservations	==0.2.1
azure-mgmt-resource	==2.2.0
azure-mgmt-scheduler	==2.0.0
azure-mgmt-search	==2.1.0
azure-mgmt-servicebus	==0.5.3
azure-mgmt-servicefabric	==0.2.0
azure-mgmt-signalr	==0.1.1
azure-mgmt-sql	==0.9.1
azure-mgmt-storage	==2.0.0
azure-mgmt-subscription	==0.2.0
azure-mgmt-trafficmanager==	0.50.0
azure-mgmt-web	==0.35.0
azure-nspkg	==3.0.2
azure-storage-blob	==12.12.0
azure-storage-common	==1.4.2
backend	==0.2.4.1
cryptography	==37.0.2
django-cors-headers	==3.10.1
django-filter	==21.1
django-storages	==1.12.3
djangorestframework	==3.13.1
djangorestframework-simplejwt
import-file	==1.11
msrest	==0.6.21
msrestazure	==0.6.4
psycopg2	==2.9.3
pycparser	==2.21
python-dateutil	==2.8.2
pytz	==2021.3
requests	==2.27.1
requests-oauthlib	==1.3.1
sqlparse	==0.4.2
typing-extensions	==4.0.1
urllib3==	1.26.9